refactor(web): 기존 Auth Api 제거하고 컨벤션에 맞게 작성
