+++ apps/web/src/http/server/auth/mutation.ts
@@ -1,0 +1,1 @@
+export {};
