+++ apps/web/src/http/server/next/index.ts
@@ -1,1 +1,2 @@
 export * from './image';
+export * from './getSession';
