+++ apps/web/src/pages/account/[sub]/update-password.tsx
@@ -2,8 +2,8 @@ import { Container, Title } from '@supercarmarket/ui';
 import type { NextPageWithLayout, Params } from '@supercarmarket/types/base';
 import layout from 'components/layout';
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
-import { getSession } from 'http/server/auth/user';
 import { AccountPasswordForm } from 'components/account';
+import { getSession } from 'http/server/next';
 
 const ProfilePasswordUpdate: NextPageWithLayout = ({
   sub,
