+++ apps/web/src/pages/community/[subject]/[category]/[id]/update.tsx
@@ -6,7 +6,7 @@ import Advertisement from 'components/common/advertisement';
 import { CommunityForm } from 'components/community';
 import { ErrorFallback } from 'components/fallback';
 import Layout from 'components/layout';
-import { getSession } from 'http/server/auth/user';
+import { getSession } from 'http/server/next';
 import type { GetServerSideProps, InferGetServerSidePropsType } from 'next';
 import { ErrorBoundary } from 'react-error-boundary';
 import { css } from 'styled-components';
